0-package.lisp,2-0-1-node.lisp: Switch to (handler-bind) in (with-cudd-critical-section);
Add (handler-bind-case) helper

[bugfix] Possible big bugfix--I think I may have misunderstood (handler-bind); Docs; finalizer bugfix(?):

Every node needs a finalizer, if only to save a reference to the parent manager,
which will keep it from being GC'd too early.  Skipping finalization entirely
when REF=NIL was probably a bug, I think.  Instead, the REF behavior is moved
into the finalizer itself, so that (cudd-recursive-deref) is called only for
nodes which incremented ref counts when created.

2-0-1-node.lisp: longer log messages when consistency checks are enabled

2-0-1-node.lisp,2-config.lisp: [WIP] refactoring config/debug-consistency-checks

deletia

[bugfix]

* Move error catcher
* formatting

* Add some :debu8 log messages, tracking where in the finalizer the trace is:
  * minor log-lvl change
  * cosmetic

* Factor out `cudd-node-logger'

* Don't pass references to CUDD elements to :log4cl macros