2-0-1-node.lisp: Finalizer bugfix(?):
Every node needs a finalizer, if only to save a reference to the parent manager,
which will keep it from being GC'd too early.  Skipping finalization entirely
when REF=NIL was probably a bug, I think.  Instead, the REF behavior is moved
into the finalizer itself, so that (cudd-recursive-deref) is called only for
nodes which incremented ref counts when created.